% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_rankings_history.R
\name{aggregate_rankings_histories}
\alias{aggregate_rankings_histories}
\title{Scrape and aggregate player rankings histories}
\usage{
aggregate_rankings_histories(x)
}
\arguments{
\item{x}{a player data frame containing player name, player rank, and profile slug.}
}
\value{
A tibble containing the historical rankings of players provided.
}
\description{
Scrape and aggregate player rankings histories
}
