% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_rankings.R
\name{get_big_player_table}
\alias{get_big_player_table}
\title{Get the big player table}
\usage{
get_big_player_table(top = NULL, category = NULL)
}
\arguments{
\item{top}{integer indicating the number of top PSA players by rank to return.}

\item{category}{character string indicating the competition category. Must be one of "both", "mens", or "womens".}
}
\value{
Tibble of players from 'http://www.squashinfo.com/players', filtered by rank and category.
}
\description{
Get the big player table
}
