% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matchup.R
\name{get_match_game_data}
\alias{get_match_game_data}
\title{Given recent match data, return a cleaned list of recent match and recent game data}
\usage{
get_match_game_data(data = NULL, players = NULL, ranks = NULL)
}
\arguments{
\item{data}{recent match data frame.}

\item{players}{a vector containing player names as strings.}

\item{ranks}{a vector of integers representing the player ranks.}
}
\value{
list containing recent match and recent games data frames.
}
\description{
Given recent match data, return a cleaned list of recent match and recent game data
}
