% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matchup.R
\name{get_matchup_profiles}
\alias{get_matchup_profiles}
\title{Get profile urls for players in matchup}
\usage{
get_matchup_profiles(
  player_1 = NULL,
  player_2 = NULL,
  players = NULL,
  ranks = NULL,
  category = NULL
)
}
\arguments{
\item{player_1}{character string of the first player's name.}

\item{player_2}{character string of the second player's name.}

\item{ranks}{integers indicating the rank of the PSA players to return.}

\item{category}{character string indicating the competition category. Must be one of "mens" or "womens".}
}
\value{
Tibble containing player names, rank, and profile url slugs.
}
\description{
Get profile urls for players in matchup
}
