% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_recent_.R
\name{get_player_profile_table}
\alias{get_player_profile_table}
\title{Get Player Summary Tables from SquashInfo}
\usage{
get_player_profile_table(data = NULL, identifier = NULL)
}
\arguments{
\item{data}{data frame with columns for player rank (int), names (chr), and profile slugs (chr).}

\item{identifier}{unique string that used to identify the desired table.}
}
\value{
Data frame containing Event Summary Table(s).
}
\description{
Given a data frame with player ranks, names, and profile slugs, \code{get_player_profile_urls()} returns the Event Summary Table of ranked players in PSA World Tour competitions.
}
