% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_recent_.R
\name{get_player_profile_urls}
\alias{get_player_profile_urls}
\title{Get Player URLs from SquashInfo}
\usage{
get_player_profile_urls(
  player = NULL,
  rank = NULL,
  category = c("mens", "womens")
)
}
\arguments{
\item{player}{character string of player name.}

\item{rank}{single integer or vector of integers indicating the rank of the PSA player(s) to return.}

\item{category}{character string indicating the competition category. Must be one of "both", "mens", or "womens".}
}
\value{
Data frame containing player ranks, names, and profile slugs.
}
\description{
Given the player name, rank(s), and competition category, \code{get_player_profile_urls()} returns profile slugs of ranked players in PSA World Tour competitions.
}
