% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_rankings_history.R
\name{get_player_rankings_history}
\alias{get_player_rankings_history}
\title{Get recent PSA rankings from SquashInfo}
\usage{
get_player_rankings_history(player = NULL, rank = NULL, category = NULL)
}
\arguments{
\item{player}{character string of player name(s).}

\item{rank}{integer indicating the rank of the PSA player(s) to return.}

\item{category}{character string indicating the competition category. Must be one of "both", "mens", or "womens".}
}
\value{
Tibble containing the player rank, player name, player's seeding, round reached, event name, event date, event location, and event tour.
}
\description{
Given a competition category, \code{get_player_rankings_history()} returns the most recent PSA rankings table.
}
\note{
This function only returns the most recent PSA rankings table for Men's and Women's singles competitions.
}
\examples{

## Get the rankings history for the top three women's singles players
top_two <- get_player_rankings_history(rank = 1:2, category = "mens")

ggplot(top_two) +
 geom_line(aes(x = exact_date, y = rank, group = name, colour = name)) +
 scale_y_reverse()

## Get the rankings history for the top three women's singles players
top_three <- get_player_rankings_history(rank = 1:3, category = "womens")

ggplot(top_three) +
  geom_line(aes(x = exact_date, y = rank, group = name, colour = name)) +
  scale_y_reverse()

## Get the top 50 players in both men's and women's singles competitions
best <- get_player_rankings_history(rank = 1, category = "both")

ggplot(best) +
  geom_line(aes(x = exact_date, y = rank, group = name, colour = name)) +
  scale_y_reverse()

}
\references{
\url{http://www.squashinfo.com/rankings/men} \cr
    \url{http://www.squashinfo.com/rankings/women}
}
