% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_recent_matches.R
\name{get_player_recent_matches}
\alias{get_player_recent_matches}
\title{Get a player's recent matches from SquashInfo}
\usage{
get_player_recent_matches(player = NULL, rank = NULL, category = NULL)
}
\arguments{
\item{player}{character string of player name(s).}

\item{rank}{integer indicating the rank of the PSA player(s) to return.}

\item{category}{character string indicating the competition category. Must be one of "both", "mens", or "womens".}
}
\value{
Tibble containing the player rank, name, opponent, match result, games won, games lost, match time, date, tournament round, event, PSA designation, and event location.
}
\description{
Given the full name or rank of a player and the competition category, \code{get_player_recent_matches()} returns scraped recent match data for PSA ranked players.
}
\note{
This function only returns data from players ranked in the most recent PSA rankings table for Men's and Women's singles competitions.
}
\examples{

## Get Mohamed Elshorbagy's most recent match data
get_player_recent_matches(player = "Mohamed Elshorbagy", category = "men")

## Get recent match data from the 4th ranked player in Women's competitions
get_player_recent_matches(rank = 4, category = "womens")

## Get recent match data for Ali Farag and Nour El Tayeb
get_player_recent_matches(player = c("Ali Farag", "Nour El Tayeb"), category = "both")

## Get recent match data from the top 5 players in both Men's and Women's competitions
get_player_recent_matches(rank = 1:5, category = "both")

}
\references{
\url{http://www.squashinfo.com/rankings/men} \cr
    \url{http://www.squashinfo.com/rankings/women}
}
