% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_players.R
\name{get_players}
\alias{get_players}
\title{Get PSA Squash Player Data from SquashInfo}
\usage{
get_players(top = 25, category = c("both", "men", "women"))
}
\arguments{
\item{top}{integer indicating the number of top PSA players to return}

\item{category}{character string indicating the competition category}
}
\value{
Tibble containing first name, last name, age, gender, birthplace, nationality, residence, height in cm, weight in kg, plays (handedness), racket brand, year of joining PSA, university, and club.
}
\description{
Returns scraped profile data men and/or women players in PSA World Tour competitions
}
\note{
This function only returns players ranked in the most recent PSA rankings table for Men's and Women's singles competitions.
}
\examples{

get_players(top = 25, category = "women")

both <- get_players(5, "both")

}
\references{
\url{http://www.squashinfo.com/rankings/men}
    \url{http://www.squashinfo.com/rankings/women}
}
