% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rankings.R
\name{get_rankings_table}
\alias{get_rankings_table}
\title{Get the PSA rankings tables from SquashInfo}
\usage{
get_rankings_table(top = NULL, category = NULL)
}
\arguments{
\item{top}{integer indicating the number of top PSA players by rank to return.}

\item{category}{character string indicating the competition category. Must be one of "mens", or "womens".}
}
\value{
Tibble containing the player rank, previous month's rank, name, highest ranking achieved, date of highest ranking, nationality, and competition category.
}
\description{
Given a competition category, \code{get_rankings_table()} returns the most recent PSA rankings table.
}
