% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matchup.R
\name{get_recent_matches}
\alias{get_recent_matches}
\title{Get a player's recent matches}
\usage{
get_recent_matches(profile_urls)
}
\arguments{
\item{profile_urls}{a data frame containing a list of player names, ranks, and profile slugs.}
}
\value{
Tibble containing the player's recent matches.
}
\description{
Get a player's recent matches
}
