% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tournament_.R
\name{get_tournament_}
\alias{get_tournament_}
\title{Scrape tournament metadata}
\usage{
get_tournament_(tournament = NULL, year = NULL, world_tour = NULL)
}
\arguments{
\item{tournament}{the name of the tournament.}

\item{year}{the year the tournament was held.}

\item{world_tour}{logical indicating whether the tournament was a part of the PSA World Tour.}
}
\value{
a data frame containing a specified tournament and its slug.
}
\description{
Scrape tournament metadata
}
