% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tournament_.R
\name{get_tournament_objects}
\alias{get_tournament_objects}
\title{Get players, matches, or games from a given tournament}
\usage{
get_tournament_objects(tournaments = NULL, level = NULL)
}
\arguments{
\item{tournaments}{a data frame containing a set of tournaments}

\item{level}{the level of object to return. Must be one of "players", "matches", or "games".}
}
\value{
a tibble containing the tournament objects.
}
\description{
Get players, matches, or games from a given tournament
}
