% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squashinformr.R
\docType{package}
\name{squashinformr}
\alias{squashinformr}
\title{\code{squashinformr} package}
\description{
The squashinformr package allows sports enthusiasts to easily access data related to the Professional Squash Association's World Tour and other squash tournaments from SquashInfo. The functions within this package scrape, parse, and clean data associated with players, tournaments, and rankings data.
}
\details{
See the README on \href{https://github.com/HaydenMacDonald/squashinformr}{GitHub}
}
