% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squashinformr.R
\docType{package}
\name{squashinformr}
\alias{squashinformr}
\title{\code{squashinformr} package}
\description{
Google spreadsheets R API
}
\details{
See the README on \href{https://github.com/HaydenMacDonald/squashinformr}{GitHub}
}
